@@ -1,7 +1,14 @@
 import { defineConfig } from 'vite'
-import reactRefresh from '@vitejs/plugin-react-refresh'
+import react from '@vitejs/plugin-react'
+import tsconfigPaths from 'vite-tsconfig-paths'
 
-// https://vitejs.dev/config/
+// https://vitejs.dev/config https://vitest.dev/config
 export default defineConfig({
-  plugins: [reactRefresh()]
+  plugins: [react(), tsconfigPaths()],
+  test: {
+    globals: true,
+    environment: 'happy-dom',
+    setupFiles: '.vitest/setup',
+    include: ['**/test.{ts,tsx}']
+  }
 })
